fixed case in which several redirections in the pipeline would fail due to incorrect stopping condition in find_op
