+++ Makefile
@@ -6,7 +6,7 @@
 #    By: cdomet-d <cdomet-d@student.42lyon.fr>      +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: Invalid date        by                   #+#    #+#              #
-#    Updated: 2024/04/16 14:52:17 by cdomet-d         ###   ########lyon.fr    #
+#    Updated: 2024/04/17 12:07:29 by cdomet-d         ###   ########lyon.fr    #
 #                                                                              #
 # **************************************************************************** #
 
@@ -75,7 +75,8 @@ SRC_BUILTIN:=	export.c \
 SRC +=  $(addprefix $(DIR_UTILS), $(SRC_UTILS))
 DIR_UTILS:= utils/
 SRC_UTILS:=		display_data.c \
-				display_structs.c \
+				display_structs_q.c \
+				display_structs_v.c \
 				error_handling.c \
 				main_utils.c \
