+++ src/exec/heredoc.c
@@ -6,7 +6,7 @@
 /*   By: cdomet-d <cdomet-d@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 11:51:17 by cdomet-d          #+#    #+#             */
-/*   Updated: 2024/04/16 15:58:49 by cdomet-d         ###   ########lyon.fr   */
+/*   Updated: 2024/04/17 12:06:03 by cdomet-d         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static void	*h_gnl(int fd, t_input *in)
 {
-	char *line;
+	char	*line;
 
 	fprintf(stderr, "%.20s\n", "-- h_gnl -----------------------------");
 	fprintf(stderr, "\033[2mdelim : [%s]\033[0m\n", in->data[0]);
@@ -45,21 +45,21 @@ static char	*gen_filename(int fn)
 	char	*strfn;
 	char	*filename;
 
-	fprintf(stderr, "%.20s\n", "-- gen_filename -----------------------------");
+	// fprintf(stderr, "%.20s\n", "-- gen_filename -----------------------------");
 	strfn = ft_itoa(fn);
 	if (!strfn)
 		return (print_error(errno, "gen_filename (itoa fn)"));
 	filename = ft_strjoin("tmp_", strfn);
 	if (!filename)
 		return (print_error(errno, "gen_filename (strjoin filename)"));
-	fprintf(stderr, "\033[2mfile [%.20s]\033[0m\n", filename);
+	// fprintf(stderr, "\033[2mfile [%.20s]\033[0m\n", filename);
 	free(strfn);
 	return (filename);
 }
 
 static void	*create_hfile(t_fd *fd, t_input *tmp, char *filename)
 {
-	fprintf(stderr, "%.20s\n", "-- create_hfile -----------------------------");
+	// fprintf(stderr, "%.20s\n", "-- create_hfile -----------------------------");
 	fd->hfd = open(filename, O_CREAT | O_TRUNC | O_RDWR, 0777);
 	if (fd->hfd == -1)
 		return (print_error(errno, "create_hfile (opening tmp)"));
@@ -88,5 +88,6 @@ void	*create_hdocs(t_fd *fd, t_input *in)
 		tmp = find_tok(tmp, heredoc, true);
 		fn++;
 	}
+	pmin(in, "in create_hdocs");
 	return ((int *) true);
 }
